Create Transaction
- Fix bugs

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,7 +19,7 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->float('uang');
+            $table->double('uang');
             $table->string('alamat');
             $table->rememberToken();
             $table->timestamps();
